Fix display of group ingredients
Commit 043f6081452365daaa033c58e0738527ccb64c3d broke a couple of things
in the course of its refactoring.  This patch restores the "G" flag that
signals group ingredients, the preference for "default:" items as group
representatives (where a representative isn't specifically registered),
and the built-in registration of default:cobble as representative for
group:stone (because it's the most commonly-used item in that group).